backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-environ==0.13.0
channels==4.3.2
daphne==4.2.1
pillow==12.1.1
gunicorn==23.0.0
psycopg[binary]==3.2.9
channels-redis==4.2.1
whitenoise==7.0
